requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-barcode[images]==0.15.1
qrcode[pil]==7.4.2
requests==2.31.0
python-decouple==3.8
django-crontab==0.7.1
reportlab==4.0.8
gunicorn==21.2.0
psycopg2-binary==2.9.9